CS490/pyproject.toml
[project]
name = "26-cs490-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.6.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "piq>=0.8.0",
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "torchview>=0.2.7",
    "torchvision>=0.25.0",
    "torchviz>=0.0.3",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
